Rename preferences to have 'group' prefix
