+++ src/data/index.js
@@ -49,7 +49,7 @@ import * as PageBuilder from './page-builder';
 import * as Person from './people';
 import * as PhoneNumber from './phone-number';
 import * as PrayerRequest from './prayer-request';
-import * as Preferences from './preferences';
+import * as GroupPreferences from './group-preferences';
 import * as RockConstants from './rock-constants';
 import * as Schedule from './schedule';
 import * as Search from './search';
@@ -87,6 +87,7 @@ const data = {
   Followings,
   Group,
   GroupItem,
+  GroupPreferences,
   Interactions,
   Interfaces,
   LinkTree,
@@ -102,7 +103,6 @@ const data = {
   PersonalDevice,
   PhoneNumber,
   PrayerRequest,
-  Preferences,
   RockConstants,
   Schedule,
   Scripture,
